add ability to set defaults
